@@ -1,51 +1,53 @@
 {
-  "name": "pick-random-cli",
-  "version": "1.0.0",
-  "description": "Pick random items from a list",
-  "license": "MIT",
-  "repository": "sindresorhus/pick-random-cli",
-  "author": {
-    "name": "Sindre Sorhus",
-    "email": "sindresorhus@gmail.com",
-    "url": "sindresorhus.com"
-  },
-  "bin": {
-    "pick-random": "cli.js"
-  },
-  "engines": {
-    "node": ">=4"
-  },
-  "scripts": {
-    "test": "xo && ava"
-  },
-  "files": [
-    "cli.js"
-  ],
-  "keywords": [
-    "cli-app",
-    "cli",
-    "pick",
-    "choose",
-    "random",
-    "rand",
-    "list",
-    "raffle",
-    "item",
-    "element",
-    "decision",
-    "winner",
-    "picking",
-    "picker",
-    "choosing"
-  ],
-  "dependencies": {
-    "get-stdin": "^5.0.1",
-    "meow": "^3.7.0",
-    "pick-random": "^2.0.0"
-  },
-  "devDependencies": {
-    "ava": "*",
-    "execa": "^0.6.0",
-    "xo": "*"
-  }
+	"name": "pick-random-cli",
+	"version": "1.0.0",
+	"description": "Pick random items from a list",
+	"license": "MIT",
+	"repository": "sindresorhus/pick-random-cli",
+	"funding": "https://github.com/sponsors/sindresorhus",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "https://sindresorhus.com"
+	},
+	"type": "module",
+	"bin": {
+		"pick-random": "./cli.js"
+	},
+	"engines": {
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
+	},
+	"scripts": {
+		"test": "xo && ava"
+	},
+	"files": [
+		"cli.js"
+	],
+	"keywords": [
+		"cli-app",
+		"cli",
+		"pick",
+		"choose",
+		"random",
+		"rand",
+		"list",
+		"raffle",
+		"item",
+		"element",
+		"decision",
+		"winner",
+		"picking",
+		"picker",
+		"choosing"
+	],
+	"dependencies": {
+		"get-stdin": "^9.0.0",
+		"meow": "^10.1.1",
+		"pick-random": "^4.0.0"
+	},
+	"devDependencies": {
+		"ava": "^3.15.0",
+		"execa": "^5.1.1",
+		"xo": "^0.44.0"
+	}
 }
